Move Equihash verification APIs into equihash crate
