+++ components/equihash/Cargo.toml
@@ -9,3 +9,6 @@ license = "MIT OR Apache-2.0"
 edition = "2018"
 
 [dependencies]
+blake2b_simd = "0.5"
+byteorder = "1"
+log = "0.4"
