+++ components/equihash/src/lib.rs
@@ -1,7 +1,5 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+//! The Equihash Proof-of-Work function.
+
+mod verify;
+
+pub use verify::{is_valid_solution, is_valid_solution_iterative, is_valid_solution_recursive};
