+++ zcash_primitives/Cargo.toml
@@ -17,6 +17,7 @@ blake2b_simd = "0.5"
 blake2s_simd = "0.5"
 byteorder = "1"
 crypto_api_chachapoly = "0.2.1"
+equihash = { version = "0.0", path = "../components/equihash" }
 ff = { version = "0.6", path = "../ff" }
 fpe = "0.2"
 hex = "0.3"
