+++ zcash_primitives/src/block.rs
@@ -9,7 +9,7 @@ use std::ops::Deref;
 
 use crate::serialize::Vector;
 
-pub mod equihash;
+pub use equihash;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub struct BlockHash(pub [u8; 32]);
